requirements/base.txt
# Base requirements - Core dependencies needed in all environments
Django==5.2.6
asgiref==3.9.2
sqlparse==0.5.3

# Django REST Framework and JWT Authentication
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
PyJWT==2.10.1

# WebSocket support
channels==4.3.1
daphne==4.2.1
autobahn==24.4.2
Twisted==25.5.0
txaio==25.9.2

# PDF processing
PyPDF2==3.0.1

# Vector database and embeddings
chromadb==1.1.0
numpy==2.3.3
onnxruntime==1.23.0

# OpenAI integration
openai==0.26.5
aiohttp==3.12.15
httpx==0.28.1
httpcore==1.0.9

# Environment and configuration
python-decouple==3.8
python-dotenv==1.1.1

# Core utilities
requests==2.32.5
python-dateutil==2.9.0.post0
packaging==25.0
PyYAML==6.0.3

# Security and crypto
bcrypt==5.0.0
cryptography==46.0.2
pyOpenSSL==25.3.0

# Core dependencies (auto-installed with above packages)
aiohappyeyeballs==2.6.1
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
attrs==25.3.0
Automat==25.4.16
backoff==2.2.1
cachetools==6.2.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
coloredlogs==15.0.1
constantly==23.10.4
distro==1.9.0
durationpy==0.10
filelock==3.19.1
flatbuffers==25.9.23
frozenlist==1.7.0
fsspec==2025.9.0
google-auth==2.41.0
googleapis-common-protos==1.70.0
grpcio==1.75.1
h11==0.16.0
humanfriendly==10.0
hyperlink==21.0.0
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
incremental==24.7.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
markdown-it-py==4.0.0
mdurl==0.1.2
mmh3==5.2.0
mpmath==1.3.0
multidict==6.6.4
oauthlib==3.3.1
opentelemetry-api==1.37.0
opentelemetry-exporter-otlp-proto-common==1.37.0
opentelemetry-exporter-otlp-proto-grpc==1.37.0
opentelemetry-proto==1.37.0
opentelemetry-sdk==1.37.0
opentelemetry-semantic-conventions==0.58b0
orjson==3.11.3
overrides==7.7.0
posthog==5.4.0
propcache==0.3.2
protobuf==6.32.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.2
pycparser==2.23
pydantic==2.11.9
pydantic_core==2.33.2
Pygments==2.19.2
PyPika==0.48.9
referencing==0.36.2
requests-oauthlib==2.0.0
rich==14.1.0
rpds-py==0.27.1
rsa==4.9.1
service-identity==24.2.0
six==1.17.0
sniffio==1.3.1
sympy==1.14.0
tenacity==9.1.2
tokenizers==0.22.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.3.0
websocket-client==1.8.0
websockets==15.0.1
yarl==1.20.1
zipp==3.23.0
zope.interface==8.0.1